LayoutTests/http/tests/site-isolation/paste-in-iframe-expected.txt

Tests that the iframe is able to receive paste events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.data is 'onpaste'
PASS successfullyParsed is true

TEST COMPLETE
